feat: add "Run to stage" option in the repl
This is the website side of a change to the repl to allow running decaffeinate
part-way to inspect intermediate results.

One nice advantage of the pretty-printed AST/token format is that you can use
the expand/collapse feature in the right-side editor automatically.

Some future improvements not done here:
* Ideally you'd be able to cross-reference code with AST nodes by clicking
  and/or hovering, like you can do at http://astexplorer.net/ . Since code
  ranges are displayed in a standard format, we could potentially just use
  regexes or something to match that format in the repl JS code.
* Ideally we'd know when the contents in the right editor are JavaScript, and
  gray out the "evaluate" button if not.

Here's a live demo:
http://www.alangpierce.com/decaffeinate-project.org/repl/#?evaluate=false&stage=full&code=x%20%3D%20a%20%22%23%7Bb%20%2B%20c%7D%22