fix: use dark text in the "run to stage" select (#4)
It looks like a recent Chrome update made it so the text color for `select`
elements gets inherited, so it was showing white on light gray for me. This was
also happening in Firefox (on Mac) but not Safari. To fix, I just specified the
color explicitly in the CSS.

Demo: http://www.alangpierce.com/decaffeinate-project.org/repl/#?evaluate=true&stage=full&code=