fix: update browserify command for new dist structure
With the TypeScript change, we publish a dist directory with index.js as the
top-level rather than decaffeinate.js, so we need to update the browserify
command accordingly so repl publishing works right.

This meant we were publishing the old decaffeinate version that's checked
into this repo. To avoid this happening in the future, I also just got rid
of that file.